+++ src/accessories/CameraAccessory.ts
@@ -16,8 +16,10 @@ import { EufyCameraStreamingDelegate } from './streamingDelegate';
  */
 export class CameraAccessory {
   private service: Service;
+  private MotionService: Service;
   private switchEnabledService: Service;
   private switchMotionService: Service;
+  private motion_triggered: boolean;
 
   constructor(
     private readonly platform: EufySecurityPlatform,
@@ -26,10 +28,14 @@ export class CameraAccessory {
   ) {
 
     this.service = {} as Service;
+    this.MotionService = {} as Service;
     this.switchEnabledService = {} as Service;
     this.switchMotionService = {} as Service;
 
     this.platform.log.debug(this.accessory.displayName, 'Constructed Camera');
+
+    this.motion_triggered = false;
+
     // set accessory information
     this.accessory
       .getService(this.platform.Service.AccessoryInformation)!
@@ -51,31 +57,22 @@ export class CameraAccessory {
         eufyDevice.getHardwareVersion(),
       );
 
-    this.service =
-      this.accessory.getService(this.platform.Service.MotionSensor) ||
-      this.accessory.addService(this.platform.Service.MotionSensor);
 
-    this.service.setCharacteristic(
-      this.platform.Characteristic.Name,
-      accessory.displayName,
-    );
 
-    // create handlers for required characteristics of Motion Sensor
-    this.service
-      .getCharacteristic(this.platform.Characteristic.MotionDetected)
-      .on('get', this.handleMotionDetectedGet.bind(this));
+    if(this.platform.config.enableCamera) {
+      this.platform.log.info(this.accessory.displayName, 'has a camera');
+      this.service = this.cameraFunction(accessory);
+      this.MotionService = this.motionFunction(accessory);
 
-    this.eufyDevice.on('motion detected', (device: Device, open: boolean) =>
-      this.onDeviceMotionDetectedPushNotification(device, open),
-    );
+      //video stream (work in progress)
+    
+      const delegate = new EufyCameraStreamingDelegate(this.platform, this.eufyDevice);
+      accessory.configureController(delegate.controller);
 
-    this.eufyDevice.on('person detected', (device: Device, open: boolean) =>
-      this.onDeviceMotionDetectedPushNotification(device, open),
-    );
-
-    this.eufyDevice.on('pet detected', (device: Device, open: boolean) =>
-      this.onDeviceMotionDetectedPushNotification(device, open),
-    );
+    } else {
+      this.platform.log.info(this.accessory.displayName, 'has a motion sensor.');
+      this.service = this.motionFunction(accessory);
+    }
 
     if(this.eufyDevice.hasBattery && this.eufyDevice.hasBattery()) {
       this.platform.log.debug(this.accessory.displayName, 'has a battery, so append batteryService characteristic to him.');
@@ -96,15 +93,19 @@ export class CameraAccessory {
         .on('get', this.handleBatteryLevelGet.bind(this));
     }
 
-    // create a new Switch service
-    this.switchEnabledService = 
-      this.accessory.getService('Enabled') ||
-      this.accessory.addService(this.platform.Service.Switch, 'Enabled', 'enabled');
-    
-    // create handlers for required characteristics
-    this.switchEnabledService.getCharacteristic(this.platform.Characteristic.On)
-      .on('get', this.handleOnGet.bind(this))
-      .on('set', this.handleOnSet.bind(this));
+    if(this.eufyDevice.isEnabled && !this.eufyDevice.isEnabled()) {
+
+      // create a new Switch service
+      this.switchEnabledService = 
+        this.accessory.getService('Enabled') ||
+        this.accessory.addService(this.platform.Service.Switch, 'Enabled', 'enabled');
+      
+      // create handlers for required characteristics
+      this.switchEnabledService.getCharacteristic(this.platform.Characteristic.On)
+        .on('get', this.handleOnGet.bind(this))
+        .on('set', this.handleOnSet.bind(this));
+
+    }
 
     if(this.eufyDevice.isIndoorCamera && !this.eufyDevice.isIndoorCamera()) {
 
@@ -129,13 +130,110 @@ export class CameraAccessory {
     }
   }
 
+  handleEventSnapshotsActiveGet(callback) {
+    this.platform.log.debug(this.accessory.displayName, 'Triggered GET EventSnapshotsActive');
+
+    // set this to a valid value for EventSnapshotsActive
+    const currentValue = this.platform.Characteristic.EventSnapshotsActive.DISABLE;
+
+    callback(null, currentValue);
+  }
+
+  /**
+   * Handle requests to set the "Event Snapshots Active" characteristic
+   */
+  handleEventSnapshotsActiveSet(value) {
+    this.platform.log.debug(this.accessory.displayName, 'Triggered SET EventSnapshotsActive:', value);
+  }
+
+  /**
+   * Handle requests to get the current value of the "HomeKit Camera Active" characteristic
+   */
+  handleHomeKitCameraActiveGet(callback) {
+    this.platform.log.debug(this.accessory.displayName, 'Triggered GET HomeKitCameraActive');
+
+    // set this to a valid value for HomeKitCameraActive
+    const currentValue = this.platform.Characteristic.HomeKitCameraActive.OFF;
+
+    callback(null, currentValue);
+  }
+
+  /**
+   * Handle requests to set the "HomeKit Camera Active" characteristic
+   */
+  handleHomeKitCameraActiveSet(value) {
+    this.platform.log.debug(this.accessory.displayName, 'Triggered SET HomeKitCameraActive:', value);
+  }
+
+  private cameraFunction(
+      accessory: PlatformAccessory,
+  ): Service {
+    const service =
+      this.accessory.getService(this.platform.Service.CameraOperatingMode) ||
+      this.accessory.addService(this.platform.Service.CameraOperatingMode);
+
+    service.setCharacteristic(
+      this.platform.Characteristic.Name,
+      accessory.displayName,
+    );
+
+    service
+      .getCharacteristic(this.platform.Characteristic.EventSnapshotsActive)
+      .on('get', this.handleEventSnapshotsActiveGet.bind(this));
+    service
+      .getCharacteristic(this.platform.Characteristic.EventSnapshotsActive)
+      .on('set', this.handleEventSnapshotsActiveSet.bind(this));
+
+    service
+      .getCharacteristic(this.platform.Characteristic.HomeKitCameraActive)
+      .on('get', this.handleHomeKitCameraActiveGet.bind(this));
+    service
+      .getCharacteristic(this.platform.Characteristic.HomeKitCameraActive)
+      .on('set', this.handleHomeKitCameraActiveSet.bind(this));
+
+    return service as Service;
+  }
+
+  private motionFunction(
+    accessory: PlatformAccessory,
+  ): Service {
+    const service =
+      this.accessory.getService(this.platform.Service.MotionSensor) ||
+      this.accessory.addService(this.platform.Service.MotionSensor);
+
+    service.setCharacteristic(
+      this.platform.Characteristic.Name,
+      accessory.displayName,
+    );
+
+    // create handlers for required characteristics of Motion Sensor
+    service
+      .getCharacteristic(this.platform.Characteristic.MotionDetected)
+      .on('get', this.handleMotionDetectedGet.bind(this));
+
+    this.eufyDevice.on('motion detected', (device: Device, motion: boolean) =>
+      this.onDeviceMotionDetectedPushNotification(device, motion),
+    );
+
+    this.eufyDevice.on('person detected', (device: Device, motion: boolean) =>
+      this.onDeviceMotionDetectedPushNotification(device, motion),
+    );
+
+    this.eufyDevice.on('pet detected', (device: Device, motion: boolean) =>
+      this.onDeviceMotionDetectedPushNotification(device, motion),
+    );
+
+    return service as Service;
+  }
+
+
   private handleRawPropertyChange(
     device: Device, 
     type: number, 
     value: string, 
     modified: number,
   ): void {
-    this.platform.log.debug(
+    this.platform.log.info(
       'Handle Camera Raw Property Changes:  -- ',
       type, 
       value, 
@@ -148,7 +246,7 @@ export class CameraAccessory {
     name: string, 
     value: PropertyValue,
   ): void {
-    this.platform.log.debug(
+    this.platform.log.info(
       'Handle Camera Property Changes:  -- ',
       name, 
       value,
@@ -186,18 +284,21 @@ export class CameraAccessory {
     this.platform.log.debug(this.accessory.displayName, 'Triggered GET MotionDetected');
 
     const currentValue = await this.isMotionDetected();
-    this.platform.log.debug(this.accessory.displayName, 'Handle Motion Sensor:  -- ', currentValue);
+    this.platform.log.debug(this.accessory.displayName, 'Handle Camera Motion:  -- ', currentValue);
 
     callback(null, currentValue as boolean);
   }
 
   private onDeviceMotionDetectedPushNotification(
     device: Device,
-    open: boolean,
+    motion: boolean,
   ): void {
+    this.platform.log.debug(this.accessory.displayName, 'Handle DoorBell motion:  -- ', motion);
+    this.motion_triggered = (this.motion_triggered) ? false : true;
+    this.platform.log.debug(this.accessory.displayName, 'Handle DoorBell motion:  -- ', this.motion_triggered);
     this.service
       .getCharacteristic(this.platform.Characteristic.MotionDetected)
-      .updateValue(open);
+      .updateValue(this.motion_triggered);
   }
 
   /**